Move UserSpaceMemoryRegion struct and related functions to memory.go
